Fix some deadlocks and races with messages sent just after connect
- The whenever block for $!connection.inited wouldn't get called after
reconnects
- There would sometimes be a deadlock when sending /cmd roominfo on
|init|
- /cmd roominfo was sometimes getting called before the bot had logged
in, meaning no response was received

This adds Connection.logged-in, which is sent to when |userupdate| is
received and the username matches the configured username. This isn't
fool-proof since force-renaming the bot and having the bot admin make it
log back into the configured account using $eval will make the whenever
block listening on Connection.logged-in run again, but that's either the
bot owner's fault for picking an inappropriate name or the server's
staff's fault for being abusive.