Refactor connection initialization and state propagation
The old code was buggy and pretty badly written. This fixes the order
certain parts of the connection initialization process are done in,
moves some connection state that didn't belong there to
PSBot::StateManager, and abstracts out all the calls to $*SCHEDULER.cue
so the code is easier to understand and write.

Sadly, updates to connection throttling had to be removed since they
didn't work.