Parse rules before commands
PS used to be very finicky with whether or not it'd respond to /cmd
queries, preventing PSBot from propagating state and requiring it to
parse commands before rules so the eval command could be used to resend
them. This is no longer the case, so parse rules before commands so the
bot can moderate its own commands if set up to do so.