Fix state propagation for real this time, redesign the games API
PSBot was never creating new users and sending `/cmd userdetails` for
each new user it finds in the userlist included in the response to `/cmd
roominfo`. This meant that some users were never having their state
propagated.

PSBot::StateManager.propagated no longer exists; it is split into
users-propagated and rooms-propagated promises instead. Since these are
redefined on each reconnect, the react block in PSBot.start wasn't
detecting when state gets propagated again, so wrap it in a loop and
leave once the bot disconnects.

The games API is no longer codependent on users and rooms. Rooms can now
have any number of games running in them at the same time, and likewise,
users can be in as many games as they want. This makes more types of
"games" possible to implement.