Send /cmd userdetails more sparingly
Rather than sending it every time the bot receives a join or rename
message, send it every time one of those messages actually changes the
shape of the bot's state instead. This makes it so the bot only sends it
once when receiving a bunch of join or rename messages for the same user
from different rooms.