Mitigate a bug in user renames, handle sending /cmd concurrently
Users sometimes don't exist with either their old userid or their new
one. This is bad, but I can't tell what's causing this at the moment, so
ignore when this happens apart from logging it.

Sending thousands of messages at once on connect in parallel was kind of
a bad idea. It's much faster to do concurrently.